# docs/public/api/openapi-3.0.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  version: "1.0"
paths:
  /v1/actions/analyze-photo:
    post:
      security:
        - Bearer: []
      description: Identifies a physical item from a photo using the configured vision
        AI provider
      tags:
        - Actions
      summary: Analyze Item Photo
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: Photo of a single item (JPEG/PNG/WebP)
                  type: string
                  format: binary
              required:
                - file
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.AnalyzePhotoResponse"
  /v1/actions/analyze-photo-bulk:
    post:
      security:
        - Bearer: []
      description: Identifies every distinct item in a photo of an open container
        using the configured vision AI provider
      tags:
        - Actions
      summary: Analyze Container Contents Photo
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: Photo of an open container/shelf (JPEG/PNG/WebP)
                  type: string
                  format: binary
              required:
                - file
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.AnalyzeBulkResponse"
  /v1/actions/create-missing-thumbnails:
    post:
      security:
        - Bearer: []
      description: Creates thumbnails for items that are missing them
      tags:
        - Actions
      summary: Create Missing Thumbnails
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/ensure-asset-ids:
    post:
      security:
        - Bearer: []
      description: Ensures all items in the database have an asset ID
      tags:
        - Actions
      summary: Ensure Asset IDs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/ensure-import-refs:
    post:
      security:
        - Bearer: []
      description: Ensures all items in the database have an import ref
      tags:
        - Actions
      summary: Ensures Import Refs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/set-primary-photos:
    post:
      security:
        - Bearer: []
      description: Sets the first photo of each item as the primary photo
      tags:
        - Actions
      summary: Set Primary Photos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/wipe-inventory:
    post:
      security:
        - Bearer: []
      description: Deletes all items in the inventory
      tags:
        - Actions
      summary: Wipe Inventory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1.WipeInventoryOptions"
        description: Wipe options
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/zero-item-time-fields:
    post:
      security:
        - Bearer: []
      description: Resets all item date fields to the beginning of the day
      tags:
        - Actions
      summary: Zero Out Time Fields
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
...
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    repo.BarcodeProduct:
      type: object
      properties:
        barcode:
          type: string
        imageBase64:
          type: string
        imageURL:
          type: string
        item:
          $ref: "#/components/schemas/repo.EntityCreate"
        manufacturer:
          type: string
        modelNumber:
          description: Identifications
          type: string
        notes:
          description: Extras
          type: string
        search_engine_name:
          type: string
...
    repo.EntityCreate:
      type: object
      required:
        - name
      properties:
        description:
          type: string
          maxLength: 1000
        entityTypeId:
          type: string
        icon:
          type: string
          maxLength: 255
        manufacturer:
          type: string
          maxLength: 255
        modelNumber:
          description: Identifications
          type: string
          maxLength: 255
        name:
          type: string
          maxLength: 255
          minLength: 1
        parentId:
          type: string
          nullable: true
        quantity:
          type: number
        tagIds:
          description: Edges
          type: array
          items:
            type: string
...
    v1.ActionAmountResult:
      type: object
      properties:
        completed:
          type: integer
    v1.AnalyzeBulkResponse:
      type: object
      properties:
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/v1.BulkItemCandidate"
        lane:
          type: string
    v1.AnalyzePhotoResponse:
      type: object
      properties:
        categoryHints:
          type: array
          items:
            type: string
        confidence:
          type: number
        lane:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/repo.BarcodeProduct"
...
    v1.BulkItemCandidate:
      type: object
      properties:
        categoryHints:
          type: array
          items:
            type: string
        confidence:
          type: number
        description:
          type: string
        manufacturer:
          type: string
        modelNumber:
          type: string
        name:
          type: string
        quantity:
          type: number
...
    v1.WipeInventoryOptions:
      type: object
      properties:
        wipeLocations:
          type: boolean
        wipeMaintenance:
          type: boolean
        wipeTags:
          type: boolean
...